WCP-33 Base implementation for recordings DAO

+++ blackboardvc-portlet-api/src/main/java/org/jasig/portlet/blackboardvcportlet/dao/SessionRecordingDao.java
@@ -32,10 +32,8 @@ import com.elluminate.sas.RecordingLongResponse;
 public interface SessionRecordingDao {
     Set<SessionRecording> getAllRecordings();
     
-    //TODO get all recordings for user? 
-
     SessionRecording createOrUpdateRecording(RecordingLongResponse recordingLongResponse);
     
-    int deleteRecordings(int... recordingIds);
+    int deleteRecordings(long... recordingIds);
     
 }
